working game initial commit
# Guessing Game 🎉

Welcome to the Guessing Game! This is a simple and fun game where players try to guess the correct number. If they guess correctly, they are greeted with a celebratory message.

## Description

In this game, the player is prompted to guess a number. If the guess is correct, a congratulatory message is displayed. This game is a great way to practice your JavaScript skills and have some fun!

 Open `index.html` in your browser to start playing the game.

## Usage

1. Open the game in your browser.
2. Enter your guess in the input field.
3. Click the "Submit" button to see if your guess is correct.
4. If your guess is correct, you will see a celebratory message!

## Contributing

Feel free to submit issues and pull requests. We welcome contributions from everyone!

## Keywords

- Guessing Game
- JavaScript Game
- Number Guessing
- Fun Game
- Interactive Game

## License

This project is licensed under the MIT License. See the LICENSE file for details.

## Contact

For any questions or feedback, please contact korigkk@gmail.com.

---

#Guessing Game #JavaScript Game #Number Guessing and #Fun_Game

Happy coding! 🎉